requirements.txt
crewai
crewai_tools
openai>=1.7.1,<2.0.0
langchain>=0.1.0,<0.2.0
langchain-openai>=0.0.5,<0.0.6
instructor>=0.5.2,<0.6.0
numpy
regex>=2023.12.25,<2024.0.0
fastapi
uvicorn
sqlmodel
python-jose
passlib[bcrypt]